EES-3650 Move generating release content from Publisher.Services.ReleaseService to Content.Service.ReleaseService and make use of ReleaseFileRepository to get download files.
