+++ src/GovUk.Education.ExploreEducationStatistics.Content.Services.Tests/ViewModels/AllMethodologiesThemeViewModelTests.cs
@@ -1,7 +1,6 @@
 ﻿#nullable enable
 using System.Collections.Generic;
 using GovUk.Education.ExploreEducationStatistics.Content.Services.ViewModels;
-using GovUk.Education.ExploreEducationStatistics.Publisher.Model.ViewModels;
 using Xunit;
 using static GovUk.Education.ExploreEducationStatistics.Common.Services.CollectionUtils;
