+++ src/GovUk.Education.ExploreEducationStatistics.Content.Services/Interfaces/Cache/IReleaseCacheService.cs
@@ -2,7 +2,6 @@
 using System.Threading.Tasks;
 using GovUk.Education.ExploreEducationStatistics.Common.Model;
 using GovUk.Education.ExploreEducationStatistics.Content.Services.ViewModels;
-using GovUk.Education.ExploreEducationStatistics.Publisher.Model.ViewModels;
 using Microsoft.AspNetCore.Mvc;
 
 namespace GovUk.Education.ExploreEducationStatistics.Content.Services.Interfaces.Cache;
