+++ src/GovUk.Education.ExploreEducationStatistics.Content.Services/Interfaces/IReleaseService.cs
@@ -1,4 +1,5 @@
 ﻿#nullable enable
+using System;
 using System.Collections.Generic;
 using System.Threading.Tasks;
 using GovUk.Education.ExploreEducationStatistics.Common.Model;
@@ -9,6 +10,8 @@ namespace GovUk.Education.ExploreEducationStatistics.Content.Services.Interfaces
 {
     public interface IReleaseService
     {
+        Task<Either<ActionResult, CachedReleaseViewModel>> GetRelease(Guid releaseId, DateTime published);
+
         Task<Either<ActionResult, List<ReleaseSummaryViewModel>>> List(string publicationSlug);
     }
 }
