+++ src/GovUk.Education.ExploreEducationStatistics.Content.Services/ViewModels/CachedReleaseViewModel.cs
@@ -3,7 +3,7 @@ using System;
 using System.Collections.Generic;
 using GovUk.Education.ExploreEducationStatistics.Common.Model;
 
-namespace GovUk.Education.ExploreEducationStatistics.Publisher.Model.ViewModels
+namespace GovUk.Education.ExploreEducationStatistics.Content.Services.ViewModels
 {
     public record CachedReleaseViewModel
     {
