+++ src/GovUk.Education.ExploreEducationStatistics.Content.Services/ViewModels/DataBlockViewModel.cs
@@ -4,7 +4,7 @@ using GovUk.Education.ExploreEducationStatistics.Common.Model.Chart;
 using GovUk.Education.ExploreEducationStatistics.Common.Model.Data;
 using GovUk.Education.ExploreEducationStatistics.Common.Model.Data.Query;
 
-namespace GovUk.Education.ExploreEducationStatistics.Publisher.Model.ViewModels
+namespace GovUk.Education.ExploreEducationStatistics.Content.Services.ViewModels
 {
     public class DataBlockViewModel : IContentBlockViewModel
     {
