+++ src/GovUk.Education.ExploreEducationStatistics.Content.Services/ViewModels/DataGuidanceViewModel.cs
@@ -1,7 +1,6 @@
 ﻿#nullable enable
 using System.Collections.Generic;
 using GovUk.Education.ExploreEducationStatistics.Data.Services.ViewModels;
-using GovUk.Education.ExploreEducationStatistics.Publisher.Model.ViewModels;
 
 namespace GovUk.Education.ExploreEducationStatistics.Content.Services.ViewModels
 {
