+++ src/GovUk.Education.ExploreEducationStatistics.Content.Services/ViewModels/IContentBlockViewModel.cs
@@ -1,6 +1,7 @@
-﻿using System;
+﻿#nullable enable
+using System;
 
-namespace GovUk.Education.ExploreEducationStatistics.Publisher.Model.ViewModels
+namespace GovUk.Education.ExploreEducationStatistics.Content.Services.ViewModels
 {
     public interface IContentBlockViewModel
     {
