+++ src/GovUk.Education.ExploreEducationStatistics.Content.Services/ViewModels/LinkViewModel.cs
@@ -1,0 +1,13 @@
+﻿#nullable enable
+using System;
+
+namespace GovUk.Education.ExploreEducationStatistics.Content.Services.ViewModels;
+
+public class LinkViewModel
+{
+    public Guid Id { get; set; }
+
+    public string Description { get; set; } = string.Empty;
+
+    public string Url { get; set; } = string.Empty;
+}
