+++ src/GovUk.Education.ExploreEducationStatistics.Content.Services/ViewModels/PreReleaseAccessListViewModel.cs
@@ -1,6 +1,4 @@
 #nullable enable
-using GovUk.Education.ExploreEducationStatistics.Publisher.Model.ViewModels;
-
 namespace GovUk.Education.ExploreEducationStatistics.Content.Services.ViewModels
 {
     public record PreReleaseAccessListViewModel : ReleaseSummaryViewModel
