+++ src/GovUk.Education.ExploreEducationStatistics.Content.Services/ViewModels/ReleaseNoteViewModel.cs
@@ -1,0 +1,13 @@
+﻿#nullable enable
+using System;
+
+namespace GovUk.Education.ExploreEducationStatistics.Content.Services.ViewModels;
+
+public record ReleaseNoteViewModel
+{
+    public Guid Id { get; set; }
+
+    public string Reason { get; set; } = string.Empty;
+
+    public DateTime On { get; set; }
+}
