+++ src/GovUk.Education.ExploreEducationStatistics.Content.Services/ViewModels/ReleaseTypeViewModel.cs
@@ -1,7 +1,7 @@
 ﻿#nullable enable
 using System;
 
-namespace GovUk.Education.ExploreEducationStatistics.Publisher.Model.ViewModels
+namespace GovUk.Education.ExploreEducationStatistics.Content.Services.ViewModels
 {
     [Obsolete("TODO EES-3038 - Replace with ReleaseType. Requires a content cache refresh", false)]
     public record ReleaseTypeViewModel
