+++ src/GovUk.Education.ExploreEducationStatistics.Content.Services/ViewModels/ReleaseViewModel.cs
@@ -6,7 +6,6 @@ using GovUk.Education.ExploreEducationStatistics.Common.Extensions;
 using GovUk.Education.ExploreEducationStatistics.Common.Model;
 using GovUk.Education.ExploreEducationStatistics.Common.Services;
 using GovUk.Education.ExploreEducationStatistics.Content.Model;
-using GovUk.Education.ExploreEducationStatistics.Publisher.Model.ViewModels;
 using Newtonsoft.Json;
 using Newtonsoft.Json.Converters;
