--- src/GovUk.Education.ExploreEducationStatistics.Publisher.Model/ViewModels/LinkViewModel.cs
@@ -1,13 +1,0 @@
-﻿using System;
-
-namespace GovUk.Education.ExploreEducationStatistics.Publisher.Model.ViewModels
-{
-    public class LinkViewModel
-    {
-        public Guid Id { get; set; }
-
-        public string Description { get; set; }
-
-        public string Url { get; set; }
-    }
-}
