--- src/GovUk.Education.ExploreEducationStatistics.Publisher.Model/ViewModels/MarkDownBlockViewModel.cs
@@ -1,15 +1,0 @@
-﻿using System;
-
-namespace GovUk.Education.ExploreEducationStatistics.Publisher.Model.ViewModels
-{
-    public class MarkDownBlockViewModel : IContentBlockViewModel
-    {
-        public Guid Id { get; set; }
-
-        public int Order { get; set; }
-
-        public string Body { get; set; }
-
-        public string Type => "MarkDownBlock";
-    }
-}
