--- src/GovUk.Education.ExploreEducationStatistics.Publisher/MappingProfiles.cs
@@ -1,63 +1,0 @@
-using System.Linq;
-using AutoMapper;
-using GovUk.Education.ExploreEducationStatistics.Common.Extensions;
-using GovUk.Education.ExploreEducationStatistics.Content.Model;
-using GovUk.Education.ExploreEducationStatistics.Content.Services.ViewModels;
-using GovUk.Education.ExploreEducationStatistics.Publisher.Model.ViewModels;
-
-namespace GovUk.Education.ExploreEducationStatistics.Publisher
-{
-    /**
-     * AutoMapper Profile which is configured by AutoMapper.Extensions.Microsoft.DependencyInjection.
-     */
-    public class MappingProfiles : Profile
-    {
-        public MappingProfiles()
-        {
-            CreateMap<Link, LinkViewModel>();
-
-            CreateContentBlockMap();
-
-            CreateMap<ContentSection, ContentSectionViewModel>().ForMember(dest => dest.Content,
-                m => m.MapFrom(section => section.Content.OrderBy(contentBlock => contentBlock.Order)));
-
-            CreateMap<Release, CachedReleaseViewModel>()
-                .ForMember(dest => dest.CoverageTitle,
-                    m => m.MapFrom(release => release.TimePeriodCoverage.GetEnumLabel()))
-                .ForMember(dest => dest.Type,
-                    m => m.MapFrom(release => new ReleaseTypeViewModel
-                    {
-                        Title = release.Type.GetTitle()
-                    }))
-                .ForMember(
-                    dest => dest.Updates,
-                    m => m.MapFrom(r => r.Updates.OrderByDescending(update => update.On)))
-                .ForMember(
-                    dest => dest.Content,
-                    m => m.MapFrom(r => r.GenericContent.OrderBy(s => s.Order)));
-
-            CreateMap<Update, ReleaseNoteViewModel>();
-
-            CreateMap<MethodologyVersion, MethodologyVersionSummaryViewModel>();
-        }
-
-        private void CreateContentBlockMap()
-        {
-            CreateMap<ContentBlock, IContentBlockViewModel>()
-                .IncludeAllDerived();
-
-            CreateDataBlockMap();
-
-            CreateMap<HtmlBlock, HtmlBlockViewModel>();
-
-            CreateMap<MarkDownBlock, MarkDownBlockViewModel>();
-        }
-
-        private void CreateDataBlockMap()
-        {
-            CreateMap<DataBlock, DataBlockViewModel>();
-
-            CreateMap<DataBlockSummary, DataBlockSummaryViewModel>();
-        }
-    }
-}
