+++ src/GovUk.Education.ExploreEducationStatistics.Publisher/Services/Interfaces/IReleaseService.cs
@@ -4,8 +4,6 @@ using System.Collections.Generic;
 using System.Threading.Tasks;
 using GovUk.Education.ExploreEducationStatistics.Common.Model;
 using GovUk.Education.ExploreEducationStatistics.Content.Model;
-using GovUk.Education.ExploreEducationStatistics.Publisher.Model.ViewModels;
-using GovUk.Education.ExploreEducationStatistics.Publisher.Models;
 
 namespace GovUk.Education.ExploreEducationStatistics.Publisher.Services.Interfaces
 {
@@ -17,17 +15,9 @@ namespace GovUk.Education.ExploreEducationStatistics.Publisher.Services.Interfac
 
         Task<IEnumerable<Release>> GetAmendedReleases(IEnumerable<Guid> releaseIds);
 
-        Task<List<FileInfo>> GetDownloadFiles(Release release);
-
         Task<List<File>> GetFiles(Guid releaseId, params FileType[] types);
 
         Task<Release> GetLatestRelease(Guid publicationId, IEnumerable<Guid> includedReleaseIds);
-
-        Task<CachedReleaseViewModel> GetLatestReleaseViewModel(Guid publicationId,
-            IEnumerable<Guid> includedReleaseIds,
-            PublishContext context);
-
-        Task<CachedReleaseViewModel> GetReleaseViewModel(Guid id, PublishContext context);
 
         Task SetPublishedDates(Guid id, DateTime published);
